Refactor BincodeSerializer into an independent module
