+++ aerosim-data/src/lib.rs
@@ -99,14 +99,12 @@ fn _aerosim_data(py: Python, m: &Bound<'_, PyModule>) -> PyResult<()> {
 
     #[cfg(feature = "kafka")]
     middleware_module.add_class::<middleware::kafka::KafkaSerializer>()?;
-    #[cfg(feature = "kafka")]
-    middleware_module.add_class::<middleware::kafka::BincodeSerializer>()?;
     #[cfg(feature = "dds")]
     middleware_module.add_class::<middleware::dds::DDSSerializer>()?;
     #[cfg(feature = "zenoh")]
     middleware_module.add_class::<middleware::zenoh::ZenohSerializer>()?;
-    #[cfg(feature = "zenoh")]
-    middleware_module.add_class::<middleware::zenoh::BincodeSerializer>()?;
+
+    middleware_module.add_class::<middleware::serializers::bincode::BincodeSerializer>()?;
 
     // Add the types submodule to the main module
     m.add_submodule(&middleware_module)?;
