+++ aerosim-data/src/middleware/kafka.rs
@@ -29,23 +29,6 @@ use crate::{
 };
 
 #[pyclass]
-pub struct BincodeSerializer;
-
-impl Serializer for BincodeSerializer {
-    fn serializer(&self) -> SerializerEnum {
-        SerializerEnum::from(Self {})
-    }
-
-    fn serialize<T: Serialize>(&self, data: &T) -> Option<Vec<u8>> {
-        bincode::serialize(data).ok()
-    }
-
-    fn deserialize<T: for<'de> Deserialize<'de>>(&self, payload: &[u8]) -> Option<T> {
-        bincode::deserialize(payload).ok()
-    }
-}
-
-#[pyclass]
 pub struct KafkaSerializer;
 
 impl Serializer for KafkaSerializer {
@@ -68,7 +51,7 @@ pub struct KafkaMiddleware {
     admin: OnceLock<AdminClient<DefaultClientContext>>,
     producer: OnceLock<Arc<FutureProducer>>,
     consumers: Mutex<Vec<Arc<StreamConsumer>>>,
-    // Temporary producer with specific settings for image publishing.  
+    // Temporary producer with specific settings for image publishing.
     // This setup will remain in place until a configurable producer interface is provided to the user.
     image_producer: OnceLock<Arc<FutureProducer>>,
 }
@@ -159,7 +142,7 @@ impl MiddlewareRaw for KafkaMiddleware {
         //             .expect("Couldn't create Kafka producer."),
         //     )
         // }));
-        
+
         // Temporary image producer with specific settings to improve performance.
         let producer = match _message_type {
             "aerosim::types::CompressedImage" => Arc::clone(self.image_producer.get_or_init(|| {
@@ -398,78 +381,6 @@ impl KafkaMiddleware {
         tokio::task::block_in_place(|| {
             handle.block_on(self.pysubscribe_all_raw_impl(py, topics, callback))
         })
-    }
-}
-
-impl PySerializer for BincodeSerializer {}
-
-#[pymethods]
-impl BincodeSerializer {
-    #[new]
-    fn pynew(_py: Python) -> PyResult<Self> {
-        Ok(Self {})
-    }
-
-    #[pyo3(name = "serialize_message")]
-    fn pyserialize_message(
-        &self,
-        py: Python<'_>,
-        metadata: Metadata,
-        data: PyObject,
-    ) -> Option<Vec<u8>> {
-        let serializer = SerializerEnum::from(BincodeSerializer {});
-        self.pyserialize_message_impl(py, &serializer, metadata, data)
-    }
-
-    #[pyo3(name = "deserialize_message")]
-    fn pydeserialize_message(
-        &self,
-        py: Python<'_>,
-        message_type: PyObject,
-        payload: &[u8],
-    ) -> Option<(Metadata, PyObject)> {
-        let serializer = SerializerEnum::from(BincodeSerializer {});
-        self.pydeserialize_message_impl(py, &serializer, message_type, payload)
-    }
-
-    #[pyo3(name = "deserialize_metadata")]
-    fn pydeserialize_metadata(&self, py: Python<'_>, payload: &[u8]) -> Option<Metadata> {
-        let serializer = SerializerEnum::from(BincodeSerializer {});
-        self.pydeserialize_metadata_impl(py, &serializer, payload)
-    }
-
-    #[pyo3(name = "deserialize_data")]
-    fn pydeserialize_data(
-        &self,
-        py: Python<'_>,
-        message_type: PyObject,
-        payload: &[u8],
-    ) -> Option<PyObject> {
-        let serializer = SerializerEnum::from(BincodeSerializer {});
-        self.pydeserialize_data_impl(py, &serializer, message_type, payload)
-    }
-
-    #[pyo3(name = "from_json")]
-    fn pyserialize_from_json(
-        &self,
-        py: Python<'_>,
-        type_name: &str,
-        metadata: &Metadata,
-        data: PyObject,
-    ) -> Option<Vec<u8>> {
-        let serializer = SerializerEnum::from(BincodeSerializer {});
-        self.pyserialize_from_json_impl(py, &serializer, type_name, metadata, data)
-    }
-
-    #[pyo3(name = "to_json")]
-    fn pydeserialize_to_json(
-        &self,
-        py: Python<'_>,
-        type_name: &str,
-        payload: &[u8],
-    ) -> Option<PyObject> {
-        let serializer = SerializerEnum::from(BincodeSerializer {});
-        self.pydeserialize_to_json_impl(py, &serializer, type_name, payload)
     }
 }
