+++ aerosim-data/src/middleware/serializers/mod.rs
@@ -1,0 +1,1 @@
+pub mod bincode;
